--- a/export_mapof_test.go
+++ b/export_mapof_test.go
@@ -3,6 +3,10 @@
 
 package xsync
 
+type (
+	BucketOfPadded = bucketOfPadded
+)
+
 func CollectMapOfStats[K comparable, V any](m *MapOf[K, V]) MapStats {
 	return MapStats{m.stats()}
 }