* `MapOf`'s design is now quite different from the `Map`'s one. `MapOf` uses immutable entry structs, so there are less objects allocated on the heap and, thus, less GC pressure. There is also no need to store hash tops - the whole hash code is now stored in buckets. `sync.Mutex` is used instead of a in-house spin lock.
* Hand-made hash function is now used for integers. It's of mediocre quality, but the compiler inlines it more happily than `memhash64`.

Benchmarks for `MapOf[int, int]` with 1K entries. The change is very noticeable on such small maps and fixed size keys.

Before:
```bash
$ go test -run=^$ -bench IntegerMapOf_WarmUp
goos: linux
goarch: amd64
pkg: github.com/puzpuzpuz/xsync/v2
cpu: 11th Gen Intel(R) Core(TM) i7-1185G7 @ 3.00GHz
BenchmarkIntegerMapOf_WarmUp/reads=100%-8         	293524755	         4.124 ns/op
BenchmarkIntegerMapOf_WarmUp/reads=99%-8          	236453667	         5.025 ns/op
BenchmarkIntegerMapOf_WarmUp/reads=90%-reads-8    	168052021	         7.140 ns/op
BenchmarkIntegerMapOf_WarmUp/reads=75%-reads-8    	115658304	        10.30 ns/op
PASS
ok  	github.com/puzpuzpuz/xsync/v2	7.254s
```

After:
```bash
$ go test -run=^$ -bench IntegerMapOf_WarmUp
goos: linux
goarch: amd64
pkg: github.com/puzpuzpuz/xsync/v2
cpu: 11th Gen Intel(R) Core(TM) i7-1185G7 @ 3.00GHz
BenchmarkIntegerMapOf_WarmUp/reads=100%-8         	452138901	         2.735 ns/op
BenchmarkIntegerMapOf_WarmUp/reads=99%-8          	316428153	         3.561 ns/op
BenchmarkIntegerMapOf_WarmUp/reads=90%-reads-8    	246066069	         5.068 ns/op
BenchmarkIntegerMapOf_WarmUp/reads=75%-reads-8    	167377144	         7.001 ns/op
PASS
ok  	github.com/puzpuzpuz/xsync/v2	6.661s
```